+++ tests/LawnDart.TestUtilities/TestMetadataProvider.cs
@@ -33,7 +33,7 @@ public class TestMetadataProvider : IMetadataProvider
         eventMetadata.CorrelationId = commandMetadata.CorrelationId;
         eventMetadata.TenantId = commandMetadata.TenantId;
         eventMetadata.UserId = commandMetadata.UserId;
-        eventMetadata.CausationId = commandMetadata.CorrelationId;
+        eventMetadata.CausationId = commandMetadata.CausationId ?? commandMetadata.CorrelationId;
         eventMetadata.EventId = @event.Id.ToString();
         eventMetadata.Timestamp = @event.Timestamp;
         eventMetadata.TraceId = commandMetadata.TraceId;